+++ SpringbootMicroserviceOracle-dev/src/main/java/com/developing/app/accounts/accountmanaging/repository/ContractIdentifiersRepositoryExt.java
@@ -1,7 +1,7 @@
-package com.santander.digitalcore.accounts.accmanagement.repository;
+package com.developing.app.accounts.accountmanaging.repository;
 
-import com.santander.digitalcore.accounts.util.lib.core.exceptions.InternalServerErrorDarwinExceptionLeancore;
-import com.santander.digitalcore.accounts.util.lib.db.model.repository.ContractIdentifiersRepository;
+import com.java.developing.accounts.util.lib.core.exceptions.InternalServerErrorplsqlExceptionLeancore;
+import com.java.developing.accounts.util.lib.db.model.repository.ContractIdentifiersRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
@@ -30,7 +30,7 @@ public class ContractIdentifiersRepositoryExt extends ContractIdentifiersReposit
    * @param contractId The contract ID to check for existence.
    * @param conn       The database connection to use for the query.
    * @return true if the contract ID exists, false otherwise.
-   * @throws InternalServerErrorDarwinExceptionLeancore If an error occurs while accessing the database.
+   * @throws InternalServerErrorplsqlExceptionLeancore If an error occurs while accessing the database.
    */
   public boolean existsByContractId(String contractId, Connection conn) {
     String sql = """
@@ -42,7 +42,7 @@ public class ContractIdentifiersRepositoryExt extends ContractIdentifiersReposit
         """;
 
     if (conn == null) {
-      throw new InternalServerErrorDarwinExceptionLeancore("DC-COMMON-T-0001", "DATABASE_ERROR");
+      throw new InternalServerErrorplsqlExceptionLeancore("DC-COMMON-T-0001", "DATABASE_ERROR");
     }
     try (PreparedStatement ps = conn.prepareStatement(sql)) {
       ps.setString(1, contractId);
@@ -53,7 +53,7 @@ public class ContractIdentifiersRepositoryExt extends ContractIdentifiersReposit
       }
     } catch (SQLException e) {
       log.error("Error al verificar la existencia existsByContractId del CONTRACT_ID '{}': {}", contractId, e.getMessage());
-      throw new InternalServerErrorDarwinExceptionLeancore("DC-COMMON-T-0001", "DATABASE_ERROR", e);
+      throw new InternalServerErrorplsqlExceptionLeancore("DC-COMMON-T-0001", "DATABASE_ERROR", e);
     }
     return false;
   }