+++ SpringbootMicroserviceOracle-dev/CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## Release 1.0.0
 
-- First release of the microservice sgt-ap11085-accmanagement which implements an application of the type
+- First release of the microservice fcm-app-accountmanaging which implements an application of the type
 Servlet.